Add files via upload
This model plots the number of infected vs susceptible people in a population. As we can see, as time progresses, the rate of susceptible people lowers as the infected peoples rate increase because the susceptible become infected. This can lead to increase global burden of disease, which can ultimately lead to increased death rates.