Add files via upload
This illustrates compartment sizes and disease incidences. We see the number of people that are susceptible, infected and recovered. Compartmental models simplify the mathematical modelling of infectious diseases. The population is assigned to compartments with labels – for example, S, I, or R, (Susceptible, Infectious, or Recovered). People may progress between compartments. The order of the labels usually shows the flow patterns between the compartments. The disease incidence model shows the occurrence of new cases of the disease.